web_s2i/requirements.txt
slackclient
python-dateutil
django-meetup
click
configparser
dominate
Flask
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
mccabe
pycodestyle
pyflakes
visitor
WTForms
flask-mysqldb
passlib
meetup-api
slacker
PyYAML
python-dotenv
